Add a Material Design widget page (#285)
* Add a Material Design widget page

* Tweaked the material design catalog page

Tweaked the material design catalog page based on PR feedback.

* Changed file name to material-design-catalog.md

* Add a disclaimer to the material design catalog

* Add custom html for Pickers components

The date picker and time picker are implemented with functions instead
of widgets. So I created custom html for them. Specifically, a column
is added to the CSV file to indicate whether this row is standard. And
then I added a control flow to the template.