Move editor setup out from under the Android section (#401)
The editor/IDE paragraphs are currently placed as the last paragraph under the Android section/header. Thus, a developer that chooses to skip Android setup (e.g., an iOS developer) will never reach that paragraph and not be aware of the IDE option.

This simple moves it under a separate header.