Add instructions for pre-downloading binaries (#2557)
We recently changed Flutter's behaviour such that binary artifacts are
downloaded on an as-needed basis. While this is an improvement for many
users, in cases such as hermetic build environments with no network
access, or in cases where a user expects to have
intermittent/slow/expensive internet access provide instructions to
optionally pre-download binaries.