Vital update on json.md about nested classes (#2875)
* Vital update on json.md about nested classes

A super relevant subject is missing at this documentation and is so dificult to find in the web: how to set the program to generate the json properly when we have nested classes in our models. I have spent 2 days trying to figure it out when I could have killed this at the first moment when I first read this doc trying to learn how to use the json generation...

* Changes implemented

Suggested changes made.