Update sidenav.yml (#5361)
The existing gallery link points to the old integration test version, which we've kept around for regression testing. I think any customer who is wanting to explore the code ought to be pointed to the updated link, which is the source used for the running app immediately above.

Ideally we'd include a link to the repo in the running app itself, so we don't need two separate links, but I'll save that for another day.